requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
streamlit
langchain
langchain-openai
langchain-community
langgraph
chromadb
openai
python-dotenv
requests
plotly
pandas
